# backend/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: AdminService API
    version: 0.0.1
paths:
    /api/v1/article:
        post:
            tags:
                - AdminService
            operationId: AdminService_CreateArticle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.CreateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.SuccessReply'
    /api/v1/article/{article.id}:
        put:
            tags:
                - AdminService
            operationId: AdminService_UpdateArticle
            parameters:
                - name: article.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.UpdateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.SuccessReply'
    /api/v1/article/{id}:
        get:
            tags:
                - AdminService
            operationId: AdminService_GetArticleById
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.GetArticleReply'
        delete:
            tags:
                - AdminService
            operationId: AdminService_DeleteArticle
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.SuccessReply'
    /api/v1/articles:
        get:
            tags:
                - AdminService
            operationId: AdminService_ListArticle
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.ListArticleReply'
    /api/v1/articles/bytag:
        get:
            tags:
                - AdminService
            operationId: AdminService_GetArticlesByTag
            parameters:
                - name: tag
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.ListArticleReply'
    /api/v1/login:
        post:
            tags:
                - AdminService
            operationId: AdminService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.LoginReply'
    /api/v1/logout:
        post:
            tags:
                - AdminService
            operationId: AdminService_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.LogoutReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.LogoutReply'
    /api/v1/register:
        post:
            tags:
                - AdminService
            operationId: AdminService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.SuccessReply'
    /api/v1/user/{id}:
        put:
            tags:
                - AdminService
            operationId: AdminService_UpdateUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/admin.v1.UpdateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.SuccessReply'
    /api/v1/users:
        get:
            tags:
                - AdminService
            operationId: AdminService_ListUser
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/admin.v1.ListUserReply'
components:
    schemas:
        admin.v1.Article:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                summary:
                    type: string
                content_url:
                    type: string
                status:
                    type: string
                view_count:
                    type: string
                tags:
                    type: string
                imageUrl:
                    type: array
                    items:
                        type: string
        admin.v1.CreateArticleRequest:
            type: object
            properties:
                article:
                    $ref: '#/components/schemas/admin.v1.Article'
        admin.v1.GetArticleReply:
            type: object
            properties:
                article:
                    $ref: '#/components/schemas/admin.v1.Article'
        admin.v1.ListArticleReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/admin.v1.Article'
                total:
                    type: integer
                    format: int32
        admin.v1.ListUserReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/admin.v1.User'
                total:
                    type: integer
                    format: int32
        admin.v1.LoginReply:
            type: object
            properties:
                id:
                    type: string
                token:
                    type: string
        admin.v1.LoginReq:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        admin.v1.LogoutReply:
            type: object
            properties:
                id:
                    type: string
        admin.v1.LogoutReq:
            type: object
            properties:
                id:
                    type: string
        admin.v1.RegisterReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                email:
                    type: string
                nickname:
                    type: string
        admin.v1.SuccessReply:
            type: object
            properties:
                message:
                    type: string
                success:
                    type: string
        admin.v1.UpdateArticleRequest:
            type: object
            properties:
                article:
                    $ref: '#/components/schemas/admin.v1.Article'
        admin.v1.UpdateUserRequest:
            type: object
            properties:
                id:
                    type: string
                userName:
                    type: string
                nickName:
                    type: string
                password:
                    type: string
        admin.v1.User:
            type: object
            properties:
                id:
                    type: string
                user_name:
                    type: string
                password:
                    type: string
                nick_name:
                    type: string
                email:
                    type: string
                token:
                    type: string
tags:
    - name: AdminService